exchange: rename server package to network
