+++ packages/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(common)
 add_subdirectory(protos)
+add_subdirectory(network)
 add_subdirectory(ring_buffer)
-add_subdirectory(server)
