+++ packages/network/CMakeLists.txt
@@ -1,0 +1,8 @@
+add_library(network INTERFACE)
+target_include_directories(network INTERFACE include)
+target_link_libraries(
+  network
+  INTERFACE
+  spdlog::spdlog
+  exchange::protos)
+add_library(exchange::network ALIAS network)
